// stayos-api/src/core/guest-requests/domain/guest-request-type.enum.ts
export enum GuestRequestType {
  EXTRA_TOWELS = 'EXTRA_TOWELS',
  EXTRA_PILLOW = 'EXTRA_PILLOW',
  WATER_BOTTLES = 'WATER_BOTTLES',

  LAUNDRY_PICKUP = 'LAUNDRY_PICKUP',

  WAKE_UP_CALL = 'WAKE_UP_CALL',

  AIRPORT_PICKUP = 'AIRPORT_PICKUP',
  AIRPORT_DROP = 'AIRPORT_DROP',
  TAXI = 'TAXI',
  LUGGAGE_ASSISTANCE = 'LUGGAGE_ASSISTANCE',

  BABY_COT = 'BABY_COT',
  EXTRA_BED = 'EXTRA_BED',
  HAIR_DRYER = 'HAIR_DRYER',
  IRON_BOARD = 'IRON_BOARD',

  ROOM_CLEANING = 'ROOM_CLEANING',

  AC_ISSUE = 'AC_ISSUE',
  TV_ISSUE = 'TV_ISSUE',
  WIFI_ISSUE = 'WIFI_ISSUE',

  SPECIAL_DECORATION = 'SPECIAL_DECORATION',
  FLOWERS = 'FLOWERS',
  CAKE = 'CAKE',

  OTHER = 'OTHER',
}